requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.1.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fake-useragent==2.0.3
fastapi==0.115.6
fonttools==4.55.3
googletrans==4.0.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mongoengine==0.29.1
motor==3.6.1
msal==1.31.1
msal-extensions==1.2.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure_eval==0.2.3
py-mini-racer==0.6.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyee==11.1.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.9.2
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
pyzmq==26.2.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
selenium==4.28.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wheel==0.45.1
wsproto==1.2.0
zipp==3.21.0
emails
python-multipart